SA - 3.1.0.143 2/9/2018 - Fixed problem where 'Hide' and 'Minimize' commands were not working correctly when batch mode was used. - Fixed problem where Player dialog does not observe setting in 'Play' dropdown list.

+++ Src/SA/DlgAutoRecorder.cpp
@@ -800,7 +800,7 @@ BOOL CDlgAutoRecorder::Apply() {
         m_pDoc->SetNumberOfSamples(m_dwRecordSize / fmtParm.wBlockAlign);
         m_pDoc->SetRecordBandWidth(fmtParm.dwSamplesPerSec / 2);
         m_pDoc->SetSignalBandWidth(fmtParm.dwSamplesPerSec / 2);
-        m_pDoc->SetQuantization(fmtParm.wBitsPerSample);
+        m_pDoc->SetQuantization((BYTE)fmtParm.wBitsPerSample);
 
         m_pDoc->RestartAllProcesses();
         m_pDoc->InvalidateAllProcesses();
